+++ src/Accounts/Accounts/ChangeLog.md
@@ -19,7 +19,8 @@
 -->
 
 ## Upcoming Release
-* Customize display format of PSAzureRmAccount to hide secret of service principal [#14208]
+* Customized display format of PSAzureRmAccount to hide secret of service principal [#14208]
+* Added optional parameter `AuthScope` to `Connect-AzAccount` to support enhanced authentication of data plane features
 
 ## Version 2.2.8
 * Fallback to first valid context if current default context key is "Default" which is invalid